include/rdma/fi_peer.h: add foreach_unspec_addr API
Add function for the owner to update addressing information for peers
with unknown addresses.
When a message comes in from an unknown peer, the owner srx cannot match
the entry with a receive buffer posted with a directed receive since it
does not know who the message came from. To prevent unexpected messages
from unknown peers from getting stuck, a new API call is added here to
allow the peer to notify the owner when new address information is available
and for the owner to retrieve that updated address information per
receive entry.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>