prov/util: Add FI_PEER support to util counter
This commit include the following changes:

1. Add a flags field in util_cntr to store the attr.flags
from fi_cntr_open. Currently the only supported flag is
FI_PEER.
2. Extend the ofi_cntr_init to make it handle the cases
for both owner and peer. When a provider is
initializing util cntr as an owner (FI_PEER is not in flags),
it will allocate the cntr->peer_cntr object and initialize it.
When a provider is initializing util cntr as a peer,
it will just pass the peer_cntr from fi_peer_cntr_context->cntr.
3. Implement the util_peer_cntr_owner_ops and util_peer_cntr_fi_ops
The former one includes the inc and incerr ops defined in
fi_ops_cntr_owner. The latter one defines the read() ops for
progressing the peer provider ep, and set others as ENOSYS.
4. Introduce the ofi_peer_cntr_tx_inc and ofi_peer_cntr_rx_inc
util functions which is a generalized ofi_cntr_tx/rx_inc that
handles inc ops for both owner and peer.

Signed-off-by: Shi Jin <sjina@amazon.com>