prov/efa: Add new efa-direct fi_info objects
efa-direct is a new path that provides direct access to the EFA
device with minimal overhead and features. It does not support SAS
ordering, tagged messaging, atomics. It does not have  copy based
protocols and cannot support sending/receiving large message sizes
and HMEM devices that don't have p2p support.

The functionality provided by the efa-direct path is a subset of the
efa rdm path. When applications do not require  capabilities only
supported by the efa rdm path, it is more performant for them to use
the efa-direct path. Accordingly, the efa-direct fi_info objects are
returned before the efa rdm fi_info objects in the fi_getinfo call.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>